Force status reset all the time
